refactor log-log and weibull
