+++ lifelines/fitters/weibull_fitter.py
@@ -4,15 +4,13 @@ import warnings
 import numpy as np
 import pandas as pd
 import autograd.numpy as autograd_np
-from autograd import hessian, value_and_grad
-from scipy.optimize import minimize
+from autograd import elementwise_grad as egrad
 
 from scipy import stats
 from numpy.linalg import inv
-from lifelines.fitters import UnivariateFitter
+from lifelines.fitters import ParametericUnivariateFitter
 from lifelines.utils import (
     _to_array,
-    inv_normal_cdf,
     check_nans_or_infs,
     ConvergenceError,
     string_justify,
@@ -21,18 +19,7 @@ from lifelines.utils import (
 )
 
 
-def _negative_log_likelihood(lambda_rho, T, E):
-    n = T.shape[0]
-    lambda_, rho = lambda_rho
-    neg_ll = (
-        -autograd_np.log(rho * lambda_) * E.sum()
-        - (rho - 1) * (E * autograd_np.log(lambda_ * T)).sum()
-        + ((lambda_ * T) ** rho).sum()
-    )
-    return neg_ll / n
-
-
-class WeibullFitter(UnivariateFitter):
+class WeibullFitter(ParametericUnivariateFitter):
 
     r"""
     This class implements a Weibull model for univariate data. The model has parameterized
@@ -131,15 +118,17 @@ class WeibullFitter(UnivariateFitter):
         (self.lambda_, self.rho_), self._log_likelihood, self._hessian_ = self._fit_model(
             self.durations, self.event_observed, show_progress=show_progress
         )
+        self._fitted_parameters_ = np.array([self.lambda_, self.rho_])
+        self._fitted_parameters_names_ = ["lambda_", "rho_"]
+        
         self.variance_matrix_ = inv(self._hessian_)
 
         self.survival_function_ = self.survival_function_at_times(self.timeline).to_frame(name=self._label)
-
         self.hazard_ = self.hazard_at_times(self.timeline).to_frame(self._label)
-
         self.cumulative_hazard_ = self.cumulative_hazard_at_times(self.timeline).to_frame(self._label)
-
+       
         self.confidence_interval_ = self._bounds(alpha, ci_labels)
+        
         self.median_ = 1.0 / self.lambda_ * (np.log(2)) ** (1.0 / self.rho_)
 
         # estimation methods
@@ -147,143 +136,13 @@ class WeibullFitter(UnivariateFitter):
         self._predict_label = label
         self._update_docstrings()
 
+
         # plotting - Cumulative hazard takes priority.
         self.plot_cumulative_hazard = self.plot
-
+    
         return self
 
-    def _estimation_method(self, t):
-        return self.survival_function_at_times(t)
+    def _cumulative_hazard(self, params, times):
+        lambda_, rho_ = params
+        return (lambda_ * times) ** rho_
 
-    def hazard_at_times(self, times):
-        return pd.Series(self.lambda_ * self.rho_ * (self.lambda_ * times) ** (self.rho_ - 1), index=_to_array(times))
-
-    def survival_function_at_times(self, times):
-        return pd.Series(np.exp(-self.cumulative_hazard_at_times(times)), index=_to_array(times))
-
-    def cumulative_hazard_at_times(self, times):
-        lambda_, rho_ = self.lambda_, self.rho_
-        return pd.Series((lambda_ * times) ** rho_, index=_to_array(times))
-
-    def _fit_model(self, T, E, show_progress=True):
-
-        initial_values = np.ones(2)
-
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore")
-
-            results = minimize(
-                value_and_grad(_negative_log_likelihood),  # pylint: disable=no-value-for-parameter
-                initial_values,
-                jac=True,
-                method="L-BFGS-B",
-                args=(T, E),
-                bounds=((0.000001, None), (0.000001, None)),  # to stay well away from 0.
-                options={"disp": show_progress},
-            )
-
-            if results.success:
-                hessian_ = hessian(_negative_log_likelihood)(results.x, T, E)  # pylint: disable=no-value-for-parameter
-                return results.x, -results.fun, hessian_ * T.shape[0]
-            print(results)
-            raise ConvergenceError("Did not converge. This is a lifelines problem, not yours;")
-
-    def _bounds(self, alpha, ci_labels):
-        alpha2 = inv_normal_cdf((1.0 + alpha) / 2.0)
-        df = pd.DataFrame(index=self.timeline)
-
-        def _d_cumulative_hazard_d_lambda(lambda_, rho_, T):
-            return rho_ / lambda_ * (lambda_ * T) ** rho_
-
-        def _d_cumulative_hazard_d_rho(lambda_, rho_, T):
-            return np.log(lambda_ * T) * (lambda_ * T) ** rho_
-
-        gradient_at_mle = np.stack(
-            [
-                _d_cumulative_hazard_d_lambda(self.lambda_, self.rho_, self.timeline),
-                _d_cumulative_hazard_d_rho(self.lambda_, self.rho_, self.timeline),
-            ]
-        ).T
-
-        std_cumulative_hazard = np.sqrt(
-            np.einsum("nj,jk,nk->n", gradient_at_mle, self.variance_matrix_, gradient_at_mle)
-        )
-
-        if ci_labels is None:
-            ci_labels = ["%s_upper_%.2f" % (self._label, alpha), "%s_lower_%.2f" % (self._label, alpha)]
-        assert len(ci_labels) == 2, "ci_labels should be a length 2 array."
-
-        df[ci_labels[0]] = self.cumulative_hazard_at_times(self.timeline) + alpha2 * std_cumulative_hazard
-        df[ci_labels[1]] = self.cumulative_hazard_at_times(self.timeline) - alpha2 * std_cumulative_hazard
-        return df
-
-    def _compute_standard_errors(self):
-        var_lambda_, var_rho_ = self.variance_matrix_.diagonal()
-        return pd.DataFrame([[np.sqrt(var_lambda_), np.sqrt(var_rho_)]], index=["se"], columns=["lambda_", "rho_"])
-
-    def _compute_confidence_bounds_of_parameters(self):
-        se = self._compute_standard_errors().loc["se"]
-        alpha2 = inv_normal_cdf((1.0 + self.alpha) / 2.0)
-        return pd.DataFrame(
-            [np.array([self.lambda_, self.rho_]) + alpha2 * se, np.array([self.lambda_, self.rho_]) - alpha2 * se],
-            columns=["lambda_", "rho_"],
-            index=["upper-bound", "lower-bound"],
-        )
-
-    def _compute_z_values(self):
-        return np.asarray([self.lambda_ - 1, self.rho_ - 1]) / self._compute_standard_errors().loc["se"]
-
-    def _compute_p_values(self):
-        U = self._compute_z_values() ** 2
-        return stats.chi2.sf(U, 1)
-
-    @property
-    def summary(self):
-        """Summary statistics describing the fit.
-        Set alpha property in the object before calling.
-
-        Returns
-        -------
-        df : pd.DataFrame
-            Contains columns coef, exp(coef), se(coef), z, p, lower, upper
-        """
-        lower_upper_bounds = self._compute_confidence_bounds_of_parameters()
-        df = pd.DataFrame(index=["lambda_", "rho_"])
-        df["coef"] = [self.lambda_, self.rho_]
-        df["se(coef)"] = self._compute_standard_errors().loc["se"]
-        df["lower %.2f" % self.alpha] = lower_upper_bounds.loc["lower-bound"]
-        df["upper %.2f" % self.alpha] = lower_upper_bounds.loc["upper-bound"]
-        df["p"] = self._compute_p_values()
-        with np.warnings.catch_warnings():
-            np.warnings.filterwarnings("ignore")
-            df["-log2(p)"] = -np.log2(df["p"])
-        return df
-
-    def print_summary(self, decimals=2, **kwargs):
-        """
-        Print summary statistics describing the fit, the coefficients, and the error bounds.
-
-        Parameters
-        -----------
-        decimals: int, optional (default=2)
-            specify the number of decimal places to show
-        kwargs:
-            print additional metadata in the output (useful to provide model names, dataset names, etc.) when comparing 
-            multiple outputs. 
-
-        """
-        justify = string_justify(18)
-        print(self)
-        print("{} = {}".format(justify("number of subjects"), self.durations.shape[0]))
-        print("{} = {}".format(justify("number of events"), np.where(self.event_observed)[0].shape[0]))
-        print("{} = {:.3f}".format(justify("log-likelihood"), self._log_likelihood))
-        print("{} = {}".format(justify("hypothesis"), "lambda != 1, rho != 1"))
-
-        for k, v in kwargs.items():
-            print("{} = {}\n".format(justify(k), v))
-
-        print(end="\n")
-        print("---")
-
-        df = self.summary
-        print(df.to_string(float_format=format_floats(decimals), formatters={"p": format_p_value(decimals)}))
